% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAB_preprocessing.R
\name{create_scAB}
\alias{create_scAB}
\title{preprocess the single-cell data, bulk data, and phenotype data.}
\usage{
create_scAB(Obejct, bulk_dataset, phenotype, method = c("survival", "binary"))
}
\arguments{
\item{Obejct}{Seurat object}

\item{bulk_dataset}{matrix of bulk data}

\item{phenotype}{Phenotype data, a matrix with two columns "time" and "state", or a vector}

\item{method}{method "survival" or "binary"}
}
\value{
a scAB_data
}
\description{
preprocess the single-cell data, bulk data, and phenotype data.
}
