% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_findModule.R
\name{findModule}
\alias{findModule}
\title{identification cells above the threshold}
\usage{
findModule(H, tred = 2, do.dip = FALSE)
}
\arguments{
\item{H}{cell matrix}

\item{tred}{threshold}

\item{do.dip}{Logical, whether to do.dip}
}
\value{
a list with cells above the threshold
}
\description{
identification cells above the threshold
}
