% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAB_modeling.R
\name{NMF}
\alias{NMF}
\title{Classical non-negative matrix factorization algorithm}
\usage{
NMF(X, K, maxiter = 2000)
}
\arguments{
\item{X}{a non-negative matrix}

\item{K}{the rank of matrix factorization}

\item{maxiter}{the maximum number of iterations}
}
\value{
a list with the submatrix and loss value
}
\description{
Classical non-negative matrix factorization algorithm
}
