% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_select_alpha.R
\name{PreparePheno}
\alias{PreparePheno}
\title{Prepare Phenotype Data}
\usage{
PreparePheno(Object)
}
\arguments{
\item{Object}{A list containing analysis method and phenotype data}
}
\value{
For survival method: returns Object$phenotype as-is.
For other methods: returns a data frame with status and time columns.
}
\description{
Prepares phenotype data in the appropriate format for different analysis methods.
For survival analysis, returns the phenotype as-is. For other methods, converts
to a data frame with status and time columns.
}
\seealso{
Other scAB_optimal_param: 
\code{\link{EvaluateSingleCV}()},
\code{\link{ParallelEvaluate}()},
\code{\link{SequentialEvaluate}()},
\code{\link{select_alpha.optimized}()}
}
\concept{scAB_optimal_param}
