% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAB_preprocessing.R
\name{run_seurat}
\alias{run_seurat}
\title{run seurat preprocessing function}
\usage{
run_seurat(
  Obejct,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.method = "vst",
  dims_Neighbors = 1:40,
  dims_UMAP = 1:10,
  verbose = TRUE
)
}
\arguments{
\item{Obejct}{Obejct is a Seurat object}

\item{normalization.method}{Method for normalization.
  \itemize{
  \item LogNormalize: Feature counts for each cell are divided by the total counts for that cell and multiplied by the scale.factor.
  This is then natural-log transformed using log1p.
  \item CLR: Applies a centered log ratio transformation.
  \item RC: Relative counts. Feature counts for each cell are divided by the total counts for that cell and multiplied by the scale.factor.
  No log-transformation is applied. For counts per million (CPM) set \code{scale.factor = 1e6}.
}}

\item{scale.factor}{Sets the scale factor for cell-level normalization.}

\item{selection.method}{How to choose top variable features. Choose one of :
\itemize{
\item vst: First, fits a line to the relationship of log(variance) and log(mean) using local polynomial regression (loess).
Then standardizes the feature values using the observed mean and expected variance (given by the fitted line).
Feature variance is then calculated on the standardized values after clipping to a maximum (see clip.max parameter).
\item mean.var.plot (mvp): First, uses a function to calculate average expression (mean.function) and dispersion (dispersion.function)
for each feature. Next, divides features into num.bin (deafult 20) bins based on their average expression, and calculates
z-scores for dispersion within each bin. The purpose of this is to identify variable features while controlling for the strong
relationship between variability and average expression.
\item dispersion (disp): selects the genes with the highest dispersion values
}}

\item{dims_Neighbors}{Dimensions of reduction to use as input.}

\item{dims_UMAP}{Which dimensions to use as input features for UMAP.}

\item{verbose}{Print output.}
}
\value{
a Seurat object
}
\description{
Single-cell preprocess
}
\details{
Preprocess the single-cell raw data using functions in the \code{Seurat} package

#' This function provide a simplified-version of Seurat analysis pipeline for single-cell RNA-seq data. It contains the following steps in the pipeline:
\itemize{
   \item Normalize the count data present in a given assay.
   \item Identify the variable features.
   \item Scales and centers features in the dataset.
   \item Run a PCA dimensionality reduction.
   \item Constructs a Shared Nearest Neighbor (SNN) Graph for a given dataset.
   \item Runs the Uniform Manifold Approximation and Projection (UMAP) dimensional reduction technique.
}

you can use this function to caculate x+1,then return the value of x+1.
}
