% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAB-data.r
\docType{data}
\name{sc_dataset}
\alias{sc_dataset}
\title{example data.}
\format{
A list contains two field data:
\describe{
  \item{bulk_dataset}{bulk RNA-seq data}
  \item{sc_dataset}{single-cell RNA-seq data}
  \item{phenotype}{phenotype information of bulk data samples}
  ...
}
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-1932-8}
}
\usage{
sc_dataset
}
\description{
A dataset containing single-cell RNA-seq, bulk RNA-seq data, and phenotype data
}
\keyword{datasets}
