% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAB_modeling.R
\name{scAB}
\alias{scAB}
\title{Non-negative Matrix Factorization with phenotype and cell-cell similarity regularization, for identifing phenotype-associated cell states at different resolutions.}
\usage{
scAB(Object, K, alpha = 0.005, alpha_2 = 5e-05, maxiter = 2000)
}
\arguments{
\item{Object}{a scAB_data object}

\item{K}{the rank of matrix factorization}

\item{alpha}{Coefficient of phenotype regularization}

\item{alpha_2}{Coefficient of cell-cell similarity regularization}

\item{maxiter}{the maximum number of iterations}
}
\value{
a list with the submatrix and loss value
}
\description{
Non-negative Matrix Factorization with phenotype and cell-cell similarity regularization, for identifing phenotype-associated cell states at different resolutions.
}
