% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAB_preprocessing.R
\docType{class}
\name{scAB_data-class}
\alias{scAB_data-class}
\title{The scAB_data Class}
\description{
The scAB_data object is created from a single-cell RNA-seq data, bulk RNA-seq data, and phenotype data.
Its input consists of a Seurat object, a digital matrix, and a matrix or vector providing phenotypic information. Among them, the Seurat object should contain Shared Nearest Neighbor (SNN) Graph information.
The class provides functions for data preprocessing, integrative analysis, and visualization.
}
\details{
The key slots used in the scAB_data object are described below.
}
\section{Slots}{

\describe{
\item{\code{X}}{Correlation matrix between bulk data and single-cell data (individuals should be in rows and cells in columns)}

\item{\code{S}}{Diagonal matrix for information of individuals phenotype}

\item{\code{A}}{Shared Nearest Neighbor (SNN) Graph matrix}

\item{\code{D}}{Degree Matrix of Shared Neighbor Graph}

\item{\code{L}}{Laplacian Matrix}

\item{\code{phenotype}}{Matrix of phenotype information}
}}

